python-service/sam3_endpoint/requirements.txt
# Install latest transformers from git to get SAM3 support
git+https://github.com/huggingface/transformers.git
pillow
numpy




